adds useFormCustomInputValidation
