+++ webapp/packages/core-blocks/src/FormControls/useCustomInputValidation.ts
@@ -5,31 +5,20 @@
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
  */
-import { useContext, useEffect, useRef } from 'react';
-
-import { ExecutorInterrupter } from '@cloudbeaver/core-executor';
-
+import { useEffect, useRef } from 'react';
 import { useTranslate } from '../localization/useTranslate.js';
-import { useExecutor } from '../useExecutor.js';
-import { FormContext, type IFormContext } from './FormContext.js';
 
 export interface ICustomInputValidation<TType extends HTMLInputElement | HTMLTextAreaElement = HTMLInputElement> {
   ref: React.RefObject<TType | null>;
   revalidate: () => void;
+  revalidateAndReport: (element: TType) => boolean;
 }
 
 export function useCustomInputValidation<T = void, TType extends HTMLInputElement | HTMLTextAreaElement = HTMLInputElement>(
   validation: (value: T) => string | null,
-  formContext?: IFormContext,
 ): ICustomInputValidation<TType> {
-  const reactContext = useContext(FormContext);
-  const context = formContext ?? reactContext;
   const inputRef = useRef<TType | null>(null);
   const translate = useTranslate();
-
-  if (!context) {
-    throw new Error('useCustomInputValidation must be used within a FormContext provider');
-  }
 
   function validate(element: TType): boolean {
     let value: T = undefined as unknown as T;
@@ -48,7 +37,7 @@ export function useCustomInputValidation<T = void, TType extends HTMLInputElemen
     return true;
   }
 
-  function validateAndReport(element: TType): boolean {
+  function revalidateAndReport(element: TType): boolean {
     const valid = element.validity.valid;
     const result = validate(element);
 
@@ -65,21 +54,6 @@ export function useCustomInputValidation<T = void, TType extends HTMLInputElemen
     }
   }
 
-  useExecutor({
-    executor: context?.onValidate,
-    handlers: [
-      function validationHandler(_, context) {
-        if (!inputRef.current) {
-          return;
-        }
-
-        if (!validateAndReport(inputRef.current)) {
-          ExecutorInterrupter.interrupt(context);
-        }
-      },
-    ],
-  });
-
   useEffect(() => {
     const element = inputRef.current;
     if (!element) {
@@ -89,14 +63,14 @@ export function useCustomInputValidation<T = void, TType extends HTMLInputElemen
     function handleInput(event: Event) {
       const target = event.target as TType;
       if (target.validity.valid === false) {
-        validateAndReport(target);
+        revalidateAndReport(target);
       }
     }
 
     function handleBlur(event: Event) {
       const target = event.target as TType;
       if (target.validity.valid === true) {
-        validateAndReport(target);
+        revalidateAndReport(target);
       }
     }
 
@@ -109,5 +83,5 @@ export function useCustomInputValidation<T = void, TType extends HTMLInputElemen
     };
   });
 
-  return { ref: inputRef, revalidate };
+  return { ref: inputRef, revalidate, revalidateAndReport };
 }
