+++ webapp/packages/core-blocks/src/usePasswordValidation.ts
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2024 DBeaver Corp and others
+ * Copyright (C) 2020-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,7 @@ export function usePasswordValidation(formContext?: IFormContext) {
 
     const validation = passwordPolicyService.validatePassword(value);
     return validation.isValid ? null : validation.errorMessage;
-  }, formContext);
+  });
 
   return ref;
 }
